% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decomposition_data}
\alias{decomposition_data}
\title{decomposition_data}
\format{
A dataframe with 7 variables:
\describe{
\item{earthworm_abundance}{}
\item{functional_diversity}{}
\item{CN_ratio_litter}{}
\item{decomposition}{from leaf litter}
\item{LAI_canopy}{leaf are index}
\item{herb_cover}{percentage}
\item{detritivore_abundance}{}
}
}
\usage{
decomposition_data
}
\description{
Decomposition dataset in plots with different tree functional
diversity and soil animal abundances.
}
\keyword{datasets}
