% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_psem.R
\name{plot_psem}
\alias{plot_psem}
\title{Plot psem}
\usage{
plot_psem(model, layout = "dot", render = TRUE)
}
\arguments{
\item{model}{piecewiseSEM model obtained from psem()}

\item{layout}{layout type. "dot" or "neato"}

\item{render}{if FALSE, only return object without viewing}
}
\value{
a figure or if render = FALSE, a character of the DOT code.
}
\description{
Create a diagram of piecewiseSEM model
}
