% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_data}
\alias{soil_data}
\title{soil_data}
\format{
A dataframe with 10 variables:
\describe{
\item{id}{numeric. unique identifier}
\item{country}{character. country}
\item{biome}{character. region type}
\item{latitude, longitude}{coordinates}
\item{Cmic}{soil microbial biomass}
...
}
}
\usage{
soil_data
}
\description{
Soil microbial stochiometry dataset
}
\keyword{datasets}
